+++ tests/ui/proc-macro/auxiliary/api/literal.rs
